renamed navigation components
