+++ apps/app/src/components/DesktopNavigation/index.ts
@@ -1,0 +1,3 @@
+import { DesktopNavigation } from './DesktopNavigation';
+
+export default DesktopNavigation;
