+++ apps/app/src/components/MobileNavigation/MobileNavigation.module.scss
