+++ apps/app/src/components/MobileNavigation/index.ts
@@ -1,0 +1,2 @@
+import { MobileNavigation } from './MobileNavigation';
+export { MobileNavigation };
