--- apps/app/src/components/Navbar/index.ts
@@ -1,3 +1,0 @@
-import { Navbar } from './Navbar';
-
-export default Navbar;
