--- apps/app/src/components/Navigation/index.ts
@@ -1,2 +1,0 @@
-import { Navigation } from './Navigation';
-export { Navigation };
